[EI-136] [api] refactor: src/type 폴더에서 type 관리(임시)
