+++ api/src/numerical-guidance/infrastructure/adapter/indicator-list/indicator-list.adapter.ts
@@ -15,9 +15,7 @@ export class IndicatorListAdapter implements LoadIndicatorListPort {
       .createQueryBuilder('indicator_entity')
       .getMany();
 
-    const indicatorList = IndicatorListMapper.mapEntityToDto(indicatorEntities);
-
-    const indicators = IndicatorListDto.create(indicatorList);
+    const indicators: IndicatorListDto = IndicatorListMapper.mapEntityToDto(indicatorEntities);
 
     return indicators;
   }
